requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
uvicorn==0.30.6
streamlit==1.38.0
streamlit-option-menu==0.3.13
python-dotenv==1.0.1
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6
requests==2.31.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5 
aiosignal==1.3.1 
frozenlist==1.4.1 
html5lib==1.1 
jsonpatch==1.33 
jsonpointer==3.0.0 
langchain==0.2.16 
langchain-core==0.2.38 
langchain-text-splitters==0.2.4 
langchain_ollama==0.1.3 
langsmith==0.1.110 
langchain-google-genai==1.0.10
duckduckgo-search==6.2.11 
primp-0.6.1
lxml==5.3.0 
multidict==6.0.5 
numpy==1.26.4 
ollama==0.3.2 
orjson==3.10.7 
webencodings==0.5.1 
yarl==1.9.7
langchain-google-community==1.0.8 
tenacity==8.3.0
filelock==3.16.0 
fsspec==2024.9.0 
huggingface_hub==0.24.6
pyperclip==1.9.0
streamlit-pdf-viewer==0.0.17
pypdf==4.3.1